Simplify getOptionsDict in common view
This fixes a legacy field reference to optionExtraType which would
always through an attribtue error, meaning getOptionsDict wouldn't
actually ever show the image. This also changes the option to call
the getOptionImage() member function to do the URL getting for us.

This also removes a redunant optionValue check in which pass was set
instead of any other logic.